Relocate write functions to data access layer
