+++ packages/op-coreutils/src/op_coreutils/partitioned/writehelper.py
@@ -1,20 +1,13 @@
 from dataclasses import dataclass
 from typing import Any
 
-import polars as pl
 from overrides import EnforceOverrides, override
 
-from op_coreutils import clickhouse, duckdb_local
 from op_coreutils.logger import structlog
-from op_coreutils.storage.gcs_parquet import (
-    gcs_upload_parquet,
-    local_upload_parquet,
-)
 
 from .breakout import breakout_partitions
 from .location import DataLocation
-from .dataaccess import MARKERS_DB, init_data_access
-from .marker import Marker
+from .dataaccess import init_data_access
 from .output import OutputData, ExpectedOutput, OutputPartMeta
 
 log = structlog.get_logger()
@@ -53,7 +46,7 @@ class WriteManager(EnforceOverrides):
 
         written_parts = self.write_implementation(output_data)
 
-        write_marker(
+        client.write_marker(
             data_location=self.location,
             expected_output=self.expected_output,
             written_parts=written_parts,
@@ -67,6 +60,8 @@ class ParqueWriteManager(WriteManager):
     def write_implementation(self, output_data: Any) -> list[OutputPartMeta]:
         assert isinstance(output_data, OutputData)
 
+        client = init_data_access()
+
         parts = breakout_partitions(
             df=output_data.dataframe,
             partition_cols=["chain", "dt"],
@@ -75,7 +70,7 @@ class ParqueWriteManager(WriteManager):
 
         parts_meta: list[OutputPartMeta] = []
         for part in parts:
-            write_single_part(
+            client.write_single_part(
                 location=self.location,
                 dataframe=part.df,
                 full_path=part.meta.full_path(
@@ -85,56 +80,3 @@ class ParqueWriteManager(WriteManager):
             parts_meta.append(part.meta)
 
         return parts_meta
-
-
-def write_single_part(
-    location: DataLocation,
-    dataframe: pl.DataFrame,
-    full_path: str,
-):
-    """Write a single parquet output file for a partitioned output."""
-    if location == DataLocation.GCS:
-        gcs_upload_parquet(full_path, dataframe)
-        return
-
-    elif location == DataLocation.LOCAL:
-        local_upload_parquet(
-            path=location.with_prefix(full_path),
-            df=dataframe,
-        )
-
-        return
-
-    raise NotImplementedError()
-
-
-def write_marker(
-    data_location: DataLocation,
-    expected_output: ExpectedOutput,
-    written_parts: list[OutputPartMeta],
-    markers_table: str,
-):
-    """Write marker.
-
-    Having markers allows us to quickly check completion and perform analytics
-    over previous iterations of the ingestion process.
-
-    Markers for GCS output are written to Clickhouse.
-    Markers for local output are written to DuckDB
-
-    """
-    marker = Marker(
-        expected_output=expected_output,
-        written_parts=written_parts,
-    )
-    arrow_table = marker.to_pyarrow_table()
-
-    if data_location in (DataLocation.GCS, DataLocation.BIGQUERY):
-        clickhouse.insert_arrow("OPLABS", MARKERS_DB, markers_table, arrow_table)
-        return
-
-    elif data_location == DataLocation.LOCAL:
-        duckdb_local.insert_arrow(MARKERS_DB, markers_table, arrow_table)
-        return
-
-    raise NotImplementedError()
